+++ packages/widget/src/widget/Widget.tsx
@@ -47,6 +47,7 @@ export type WidgetProps = {
   };
   routeConfig?: WidgetRouteConfig;
   filter?: ChainFilter;
+  filterOut?: ChainFilter;
   walletConnect?: WalletConnect;
   /**
    * enables sentry session replays on the widget to help with troubleshooting errors
@@ -75,6 +76,8 @@ export type WidgetProps = {
    */
   simulate?: boolean;
   disableShadowDom?: boolean;
+  ibcEurekaHighlightedAssets?: string[];
+  assetSymbolsSortedToTop?: string[];
 } & Pick<
   NewSkipClientOptions,
   | "apiUrl"