+++ packages/client/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @skip-go/client
 
+## 0.16.26
+
+### Patch Changes
+
+- d13c452: update registries
+
+## 0.16.25
+
+### Patch Changes
+
+- 7a399d5: update registries
+
+## 0.16.24
+
+### Patch Changes
+
+- 6ec520a: Add eth sepolia to gas validation bypass
+
 ## 0.16.23
 
 ### Patch Changes